Add files via upload
Create a React Application using npx create-react-app (name) in terminal
cd to (name)
Hit npm i
npm i tailwind css and @mui/material
Add the above files (in src)
Hit npm start 

It is a UI replica for a Task which has feature of adding and deleting images from the gallery and has a navigation widget. It is a interactive front-end design made using React and Tailwind.